Ship installed no-key math demo and verify it from the wheel
